### 2022-03-19 #4245 @kthyng
Hi @gdetor and thanks for your submission. It is a bit borderline on size so I am going to ping the greater editorial board to take a look and make sure it is in scope for JOSS. This process will take 1-2 weeks, thanks.
