### 2022-03-24 548f55e
Provide FreeBSD support by swapping ngspice for native OpenSPICE python simulator. Still buggy
