+++ OpenPelt/tecsim.py
@@ -1,308 +1,0 @@
-#!/usr/bin/env python3
-import numpy as np
-import PySpice.Spice.NgSpice.Shared
-import cffi
-from PySpice.Spice.Netlist import Circuit, SubCircuit
-from PySpice.Unit import u_V, u_A, u_Ohm, u_F, u_s
-import matplotlib.pylab as plt
-
-from neural_controller import neural_controller
-
-from initialize_randomness import seed_everything
-
-
-TEMP_SENSOR_SAMPLES_PER_SEC = 5.00
-SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE = 2.00
-# SIMULATION_TIME_IN_SEC = 1800.00
-SIMULATION_TIME_IN_SEC = 3000.00
-
-COLD_SIDE_NODE = 5  # 2
-HOT_SIDE_NODE = 4  # 1
-
-
-def K_to_C(T_in_C):
-    return T_in_C - 273.15
-
-
-# Detector Circuit Parameters #
-TAMB = 296.4
-RP = 1.8
-SE = 0.05292
-K_RAD = 0.34
-C_RAD = 340.00
-K_SIL = 0.143
-C_H = 2.00
-K_M = 1.768
-C_C = 2.00
-C_CONINT = 304.00
-K_CONINT = 3.1
-
-
-class NgspiceCustomSrc(PySpice.Spice.NgSpice.Shared.NgSpiceShared):
-    def __init__(self, controller_f, **kwargs):
-        # Temporary workaround:
-        # https://github.com/FabriceSalvaire/PySpice/pull/94
-        PySpice.Spice.NgSpice.Shared.ffi = cffi.FFI()
-        super().__init__(**kwargs)
-        self.controller_f = controller_f
-        self.t = []
-        self.th_actual = []
-        self.tc_actual = []
-        self.th_sensor = []
-        self.tc_sensor = []
-        self.v = []
-        self.i = []
-        self.timestep_counter = SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE
-        self.next_v = 0.00
-        self.next_i = 0.00
-
-    def send_data(self, actual_vector_values, number_of_vectors, ngspice_id):
-        if self.timestep_counter == SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE:
-            self.th_sensor.append(
-                K_to_C(actual_vector_values['V({})'.format(HOT_SIDE_NODE)].real))
-            self.tc_sensor.append(
-                K_to_C(actual_vector_values['V({})'.format(COLD_SIDE_NODE)].real))
-            self.timestep_counter = 0
-        else:
-            self.th_sensor.append(self.th_sensor[len(self.th_sensor) - 1])
-            self.tc_sensor.append(self.tc_sensor[len(self.tc_sensor) - 1])
-            self.timestep_counter += 1
-        self.th_actual.append(
-            K_to_C(actual_vector_values['V({})'.format(HOT_SIDE_NODE)].real))
-        self.tc_actual.append(
-            K_to_C(actual_vector_values['V({})'.format(COLD_SIDE_NODE)].real))
-        self.v.append(actual_vector_values['V(11)'].real)
-        self.i.append(
-            (actual_vector_values['V(13)'].real - actual_vector_values['V(12)'].real)/RP)
-        self.next_v = self.controller_f(
-            actual_vector_values['time'].real, self.th_sensor, self.tc_sensor)
-        self.next_i = self.controller_f(
-            actual_vector_values['time'].real, self.th_sensor, self.tc_sensor)
-        self.t.append(actual_vector_values['time'].real)
-        return 0
-
-    def clear(self):
-        self.t = []
-        self.th_actual = []
[…]
