### 2022-03-25 b3d75b0
Added RC circuit test
