+++ OpenPelt/__init__.py
@@ -1,4 +1,4 @@
-from .OpenPelt import tec_plant, sequencer, circular_buffer_sequencer, TECPlate, Signal, IndVar
+from .OpenPelt import tec_plant, rc_ckt_plant, sequencer, circular_buffer_sequencer, TECPlate, Signal, IndVar
 from .controller import controller, random_agent_controller, bang_bang_controller, pid_controller
 
 try:
