+++ tests/transient.py
@@ -16,7 +16,7 @@ if __name__ == "__main__":
         os.makedirs('./figs/')
     plate_select = OpenPelt.TECPlate.HOT_SIDE
     pC = OpenPelt.tec_plant("Detector",
-                            lambda t, Th_arr: 2.1@u_A,
+                            lambda t, Th_arr: 2.1,
                             OpenPelt.Signal.CURRENT,
                             plate_select)
     start_t = time.time()
