+++ OpenPelt/OpenPelt.py
@@ -38,7 +38,7 @@ DEFAULT_TEMP_SENSOR_SAMPLES_PER_SEC = 1.00
 DEFAULT_SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE = 2.00
 DEFAULT_SIMULATION_TIME_IN_SEC = 1500.00
 ROUND_DIGITS = 1
-ERR_TOL = 0.15
+ERR_TOL = 0.01
 
 FENICS_TOL = 1e-14
 CERAMIC_K = 3.8
@@ -177,6 +177,124 @@ if INCLUDE_FENICS:
     class BottomBoundary(SubDomain):
         def inside(self, x, on_boundary):
             return on_boundary and near(x[1], -0.002, FENICS_TOL)
+
+class op_amp_lib(PySpice.Spice.OpenSPICE.Shared.OpenSPICEShared):
+
+    def __init__(self,
+                 controller_f,
+                 sim_timesteps_per_sensor_sample=DEFAULT_SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE,
+                 **kwargs):
+        # Temporary workaround:
+        # https://github.com/FabriceSalvaire/PySpice/pull/94
+        PySpice.Spice.OpenSPICE.Shared.ffi = cffi.FFI()
+        super().__init__(**kwargs)
+        self.controller_f = controller_f
+        self.t = []
+        self.v = []
+        self.v_in = []
+        self.sim_timesteps_per_sensor_sample = sim_timesteps_per_sensor_sample
+        self.timestep_counter = self.sim_timesteps_per_sensor_sample
+        self.next_v = 0.00
+        OpenSPICE.set_get_vsrc(self.get_vsrc_data)
+        OpenSPICE.set_send_data(self.send_data)
+
+    def set_controller_f(self, controller_f):
+        self.controller_f = controller_f
+
+    def send_data(self, actual_vector_values, number_of_vectors, ngspice_id):
+        self.v.append(actual_vector_values['V(3)'].real)
+        self.v_in.append(actual_vector_values['V(1)'].real)
+        self.next_v = self.controller_f()
+        self.t.append(actual_vector_values['time'].real)
+        return 0
+
+    def clear(self):
+        self.t = []
+        self.v = []
+        self.timestep_counter = self.sim_timesteps_per_sensor_sample
+        self.next_v = 0.00
+        self.next_i = 0.00
+
+    def get_t(self):
+        return self.t
+
+    def get_v(self):
+        return self.v
+
+    def get_v_in(self):
+        return self.v_in
+
+    def get_vsrc_data(self, voltage, time, node, ngspice_id):
+        voltage[0] = self.next_v
+        return 0
+
+class op_amp_plant(Circuit):
+
+    def __init__(self,
+                 name,
+                 controller_f,
+                 steady_state_cycles=1000,
+                 sim_time_in_s=DEFAULT_SIMULATION_TIME_IN_SEC,
+                 sim_timesteps_per_sensor_sample=DEFAULT_SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE,
+                 temp_sensor_samples_per_s=DEFAULT_TEMP_SENSOR_SAMPLES_PER_SEC):
+        Circuit.__init__(self, name)
+        self.controller_f = controller_f
+        self.sim_time_in_s = sim_time_in_s
+        self.sim_timesteps_per_sensor_sample = sim_timesteps_per_sensor_sample
+        self.temp_sensor_samples_per_s = temp_sensor_samples_per_s
+        self.V(INPUT_SRC, '1', self.gnd, 'dc 0 external')
[…]
