### 2022-03-26 7b83237
Got all tests except fake_neural_controller_hot running on OpenSPICE
