+++ tests/op_point_voltage.py
@@ -15,9 +15,9 @@ if __name__ == "__main__":
     if not os.path.exists('./figs/'):
         os.makedirs('./figs/')
     pC = OpenPelt.tec_plant("Detector",
-                              lambda : 0.00@u_V,
+                              lambda : 0.00,
                               OpenPelt.Signal.VOLTAGE)
-    pC.characterize_plant(-5.00, 15.0, 0.1)
+    pC.characterize_plant(-5.00, 15.0, 1.0)
     pC.plot_th_tc(OpenPelt.IndVar.VOLTAGE)
     plt.savefig('./figs/{}'.format(TEST_NAME))
     data = numpy.array([pC.get_v_arr(), pC.get_th_actual()])
