+++ tests/op_point_current.py
@@ -14,9 +14,9 @@ if __name__ == "__main__":
     if not os.path.exists('./figs/'):
         os.makedirs('./figs/')
     pC = OpenPelt.tec_plant("Detector",
-                            lambda: 2.1@u_A,
+                            lambda: 2.1,
                             OpenPelt.Signal.CURRENT)
-    pC.characterize_plant(-5.00, 5.00, 0.1)
+    pC.characterize_plant(-5.00, 5.00, 0.5)
     pC.plot_th_tc(OpenPelt.IndVar.CURRENT)
     plt.savefig('./figs/{}'.format(TEST_NAME))
     data = numpy.array([pC.get_i_arr(), pC.get_th_actual()])
