### 2022-03-29 04e4549
Fixed OpenSPICE dependency issues
