### 2022-04-06 #4245 @rkurchin
👋 Hi @danaraujocr and @tpurcell90, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
