+++ paper.bib
@@ -12,7 +12,7 @@ number = {14},
 pages = {2207-2217},
 year = {2004},
 issn = {1359-4311},
-doi = {https://doi.org/10.1016/j.applthermaleng.2004.03.001},
+doi = {10.1016/j.applthermaleng.2004.03.001},
 url = {https://www.sciencedirect.com/science/article/pii/S1359431104000675},
 author = {Reiyu Chein and Guanming Huang},
 keywords = {Thermoelectric cooler, Electronic cooling, Enforced regimes, Microchannel heat sink},
@@ -27,7 +27,7 @@ number = {1},
 pages = {15-23},
 year = {2003},
 issn = {0003-2670},
-doi = {https://doi.org/10.1016/S0003-2670(03)00676-7},
+doi = {10.1016/S0003-2670(03)00676-7},
 url = {https://www.sciencedirect.com/science/article/pii/S0003267003006767},
 author = {P.A Mosier-Boss and S.H Lieberman},
 keywords = {Volatile organic compound, Raman spectroscopy, Thermoelectric cooler},
