+++ README.md
@@ -1,5 +1,7 @@
 ![](OpenPelt.png)
 
+[![DOI](https://zenodo.org/badge/468954069.svg)](https://zenodo.org/badge/latestdoi/468954069)
+
 OpenPelt is a faster-than-real-time temperature control simulation library.
 OpenPelt contains utilities for developing and verifying temperature control
 algorithms as well as a model of a thermoelectric cooler to act as the plant.
