### 2022-04-07 #4245 @danaraujocr
Sure. I am available.
I can be a reviewer.

Le mer. 6 avr. 2022 à 17:34, Rachel Kurchin ***@***.***> a
écrit :

> 👋 Hi @danaraujocr <https://github.com/danaraujocr> and @tpurcell90
> <https://github.com/tpurcell90>, would you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
