### 2022-04-07 #4306 @editorialbot
[REVIEW]: OpenPelt: Python Framework for Thermoelectric Temperature Control System Development
**Submitting author:** @gdetor (<a href="http://orcid.org/0000-0001-5891-1702">Georgios Detorakis</a>)
**Repository:** https://github.com/thejackal360/OpenPelt-Public
**Branch with paper.md** (empty if default branch): master
**Version:** beta
