### 2022-04-08 #4306 @tpurcell90
The paper is generally well written, but there are issues with the documentation that make the using the code difficult for a new user

1) The installation instructions presented in the `README` only works if one has the right version of gcc (I think it has to be 8.2.1+ because of GLIBC_2.29). This is because instead of providing the source code for the modified version of `ngspice`as a submodule or an external code, the repository includes `libngspice.so` which was compiled with gcc-9.4. This is a major limiting factor for installing the code, and I think that the ngspice library's source code should be provided, including instructions on how to compile it. Ideally this would be done automatically with `pip`, but that would be non-trivial to set up.

2) The README has a few typos in it, including links to the private repository for OpenPelt.

3) While the tests are automated, they are done with a seperate bash script that overrides `LD_LIBRARY_PATH` to include the provided `libngspice.so` file. If you have to compile ngspice from source this breaks the script. I'd recommend converting the existing testing protocol to use `pytest`

4) The provided examples and API are not enough for a newcomer to understand how to use the code. I'd include a description of the physical model that is being modeled in the examples. Improvements to the python API would also be helpful as it is currently incomplete.

I opened issues on the repository for these problems
