+++ run_tecsim.sh
@@ -32,4 +32,4 @@ fi
 run_command="./tests/$var"
 echo $run_command
 
-$run_command
+LD_LIBRARY_PATH=$(pwd)/thejackal360-ngspice/lib $run_command
