+++ build_ngspice.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env sh
+
+cd thejackal360-ngspice/ ; ./autogen.sh ; ./configure --with-ngshared --prefix=`pwd` ; make ; make install ; cd -
