### 2022-04-09 a938f2b
Fixed a bug and cleaned up TEC model documentation
