### 2022-04-09 f19263c
Fixed a bug and cleaned up TEC model documentation
