+++ OpenPelt/OpenPelt.py
@@ -199,8 +199,29 @@ class tec_plant(Circuit):
         Instantiate circuit with controller algorithm, whose function is
         specified by controller_f.
 
+        name is the name of the TEC circuit model. Please provide a string.
+
         sig_type specifies whether we're driving a voltage or a current from
         the controller.
+
+        The mid-IR detector cooler is designed using a thermoelectric cooler (TEC) sandwiched
+        between two heat sinks, one for the TEC's hot plate and one for the TEC's cold plate.
+        plate_select specifies the heat sink (either the one connected to the hot side,
+        TECPlate.HOT_SIDE, or the one connected to the cold side) whose temperature we are
+        controlling. Imagine that the sensor is connected within the heat sink specified.
+
+        _k_rad is the hot side heat sink's thermal resistance. _c_rad is the hot side heat
+        sink's thermal capacitance. Similarly, _k_conint is the cold side heat sink's
+        thermal resistance, and _c_conint is the cold side heat sink's thermal capacitance.
+        _c_h is the hot side TEC plate's thermal capacitance. _c_c is the cold side TEC plate's
+        thermal capacitance. _k_m is the internal thermal resistance between the TEC's hot plate
+        and its cold plate.
+
+        _rp is the electrical resistance of the TEC. _se is the TEC's Seebeck coefficient.
+        This accounts for the thermoelectric generation effect that occurs when the temperature
+        gradient across the TEC plates is formed.
+
+        _tamb is the ambient temperature of the surrounding environment.
 
         The timestep parameters sim_timesteps_per_sensor_sample and temp_sensor_samples_per_s
         are used to determine a suggested timestep size. However, it should be emphasized
@@ -209,6 +230,9 @@ class tec_plant(Circuit):
         reltol. Because of this, steady_state_cycles is somewhat ambiguous as well since it
         is given in timesteps.
         """
+        # TODO: Support switching the sensor position between the heat sink and the TEC plate.
+        # This would physically reflect a thermocouple sandwiched between the heat sink and
+        # the TEC plate.
         Circuit.__init__(self, name)
         self.controller_f = controller_f
         self.sig_type = sig_type
@@ -238,14 +262,15 @@ class tec_plant(Circuit):
         self.R('5', '5', '3', _k_conint@u_Ohm)
         # ELECTRICAL PELTIER MODEL
         self.V('2', '11', '13', 0.00@u_V)
-        self.R('6', '13', '12', RP@u_Ohm)
+        self.R('6', '13', '12', _rp@u_Ohm)
         self.VCVS('1', '12', self.gnd, '1', '2', voltage_gain=_se)
         # EXTERNAL SOURCE
         self.plate_select = plate_select
         self.ncs = tec_lib(self.controller_f,
                            send_data=True,
                            plate_select=self.plate_select,
-                           steady_state_cycles=steady_state_cycles)
+                           steady_state_cycles=steady_state_cycles,
+                           _rp=RP)
         if sig_type == Signal.VOLTAGE:
             self.V(INPUT_SRC, '11', self.gnd, 'dc 0 external')
         else:
@@ -515,13 +540,25 @@ class tec_lib(PySpice.Spice.NgSpice.Shared.NgSpiceShared):
                  controller_f,
                  ref=0.00,
                  steady_state_cycles=1500,
+                 _rp=RP,
                  plate_select=TECPlate.HOT_SIDE,
                  sim_timesteps_per_sensor_sample=DEFAULT_SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE,
                  **kwargs):
         """
-        Initialize tec_lib. steady_state_cycles defines the number of
-        cycles for which values must be nearly constant before considering
-        system to be in steady state.
[…]
