+++ build_ngspice.sh
@@ -1,3 +1,8 @@
 #!/usr/bin/env sh
 
-cd thejackal360-ngspice/ ; ./autogen.sh ; ./configure --with-ngshared --prefix=`pwd` ; make ; make install ; cd -
+cd thejackal360-ngspice/ || exit;
+./autogen.sh;
+./configure "--with-ngshared --prefix=$(pwd)";
+make;
+make install;
+cd - || exit
