### 2022-04-09 a5da5dc
Added initial javadoc style docstrings
