+++ OpenPelt/OpenPelt.py
@@ -63,6 +63,9 @@ K_CONINT = 3.1
 def K_to_C(T_in_K):
     """
     Convert temperature in kelvin to temperature in celsius.
+
+    @param T_in_K: temperature in kelvin
+    @return: temperature in celsius
     """
     return T_in_K - 273.15
 
@@ -70,6 +73,9 @@ def K_to_C(T_in_K):
 def C_to_K(T_in_C):
     """
     Convert temperature in celsius to temperature in kelvin.
+
+    @param T_in_C: temperature in celsius
+    @return: temperature in kelvin
     """
     return T_in_C + 273.15
 
@@ -77,6 +83,8 @@ def C_to_K(T_in_C):
 def seed_everything(seed=1234):
     """
     Set random seed.
+
+    @param seed: random seed
     """
     random.seed(seed)
     tseed = random.randint(1, 1E6)
@@ -141,6 +149,13 @@ class circular_buffer_sequencer(sequencer):
         (b) we need to set the reference there.
 
         ngspice_custom_lib is of the class type tec_lib.
+
+        @param sequence: a list represent the sequence of reference values
+        for the controller in use. The circular buffer object will cycle through
+        these values and then jump back to the beginning after hitting the end.
+        @param ngspice_custom_lib: tec_lib object used as an interface with
+        the ngspice simulator. Need to make sure reference value in ngspice_custom_lib
+        is synchronized with value in circular buffer.
         """
         self.sequence = sequence
         self.sequence_idx = 0
@@ -153,6 +168,8 @@ class circular_buffer_sequencer(sequencer):
         Get the reference value from the circular buffer sequence. Also sets
         the reference value in self.ngspice_custom_lib, intended to be of
         the class type tec_lib.
+
+        @return: reference value
         """
         if self.ngspice_custom_lib.is_steady_state():
             if self.sequence_idx == len(self.sequence) - 1:
@@ -166,6 +183,9 @@ class circular_buffer_sequencer(sequencer):
 
 if INCLUDE_FENICS:
     class BottomBoundary(SubDomain):
+        """
+        Subdomain boundary for TEC plate heat sink in Fenics.
+        """
         def inside(self, x, on_boundary):
             return on_boundary and near(x[1], -0.002, FENICS_TOL)
 
@@ -196,34 +216,59 @@ class tec_plant(Circuit):
                  sim_timesteps_per_sensor_sample=DEFAULT_SIMULATION_TIMESTEPS_PER_SENSOR_SAMPLE,
                  temp_sensor_samples_per_s=DEFAULT_TEMP_SENSOR_SAMPLES_PER_SEC):
         """
-        Instantiate circuit with controller algorithm, whose function is
+        @param name: name is the name of the TEC circuit model. Please provide a string.
+
+        @param controller_f: instantiate circuit with controller algorithm, whose function is
         specified by controller_f.
 
-        name is the name of the TEC circuit model. Please provide a string.
-
-        sig_type specifies whether we're driving a voltage or a current from
+        @param sig_type: sig_type specifies whether we're driving a voltage or a current from
         the controller.
 
-        The mid-IR detector cooler is designed using a thermoelectric cooler (TEC) sandwiched
-        between two heat sinks, one for the TEC's hot plate and one for the TEC's cold plate.
-        plate_select specifies the heat sink (either the one connected to the hot side,
-        TECPlate.HOT_SIDE, or the one connected to the cold side) whose temperature we are
-        controlling. Imagine that the sensor is connected within the heat sink specified.
+        @param plate_select: The mid-IR detector cooler is designed using a thermoelectric
+        cooler (TEC) sandwiched between two heat sinks, one for the TEC's hot plate and one
[…]
