### 2022-04-10 63b235e
More updated in the README file. Tests directory has been reorganized. run_all file runs with pytest. Separated examples directory added.
