### 2022-04-10 4f7e083
More updated in the README file. Tests directory has been reorganized. run_all file runs with pytest. Separated examples directory added.
