+++ run_tecsim.sh
@@ -2,34 +2,28 @@
 
 if [ "$1" == "basic_bang_bang" ]
 then
-    var="basic_bang_bang.py"
+    var="./tests/basic_bang_bang.py"
 elif [ "$1" == "op_point_current" ]
 then
-    var="op_point_current.py"
+    var="./tests/op_point_current.py"
 elif [ "$1" == "op_point_voltage" ]
 then
-    var="op_point_voltage.py"
+    var="./tests/op_point_voltage.py"
 elif [ "$1" == "pid_cold" ]
 then
-    var="pid_cold.py"
+    var="./tests/pid_cold.py"
 elif [ "$1" == "pid_hot" ]
 then
-    var="pid_hot.py"
+    var="./tests/pid_hot.py"
 elif [ "$1" == "transient" ]
 then
-    var="transient.py"
+    var="./tests/transient.py"
 else
     echo "Running full test suite!"
-    ./run_tecsim.sh basic_bang_bang
-    ./run_tecsim.sh op_point_current
-    ./run_tecsim.sh op_point_voltage
-    ./run_tecsim.sh pid_cold
-    ./run_tecsim.sh pid_hot
-    ./run_tecsim.sh transient
-    exit
+    var="./tests/run_all.py"
 fi
 
-run_command="./tests/$var"
+run_command="pytest $var"
 echo $run_command
 
 LD_LIBRARY_PATH=$(pwd)/thejackal360-ngspice/lib $run_command
