+++ tests/run_all.py
@@ -7,22 +7,22 @@ from pid_cold import pid_cold
 from transient import transient
 
 
-def test_bang_bang_controller(ref_temp=65.00):
+def test_bang_bang_controller(ref_temp=25.00):
     (current, heat, cool), _ = bang_bang(reference_temperatures=[ref_temp])
     ref_vals = np.ones(heat[1, 200:].shape) * ref_temp
-    assert np.allclose(heat[1, 200:], ref_vals, rtol=1, atol=1) is True
+    assert np.allclose(heat[1, 200:], ref_vals, rtol=1e-3, atol=1) is True
 
 
-def test_pid_hot(ref_temp=75.00):
+def test_pid_hot(ref_temp=50.00):
     (voltage, heat, cool), _ = pid_hot(ref_temperatures=[ref_temp])
     ref_vals = np.ones(heat[1, 200:].shape) * ref_temp
-    assert np.allclose(heat[1, 200:], ref_vals, rtol=1, atol=1) is True
+    assert np.allclose(heat[1, 200:], ref_vals, rtol=1e-3, atol=1) is True
 
 
 def test_pid_cold(ref_temp=35.00):
     (voltage, heat, cool), _ = pid_cold(ref_temperatures=[ref_temp])
-    ref_vals = np.ones(heat[1, 200:].shape) * ref_temp
-    assert np.allclose(heat[1, 200:], ref_vals, rtol=1, atol=1) is True
+    ref_vals = np.ones(cool[1, 200:].shape) * ref_temp
+    assert np.allclose(cool[1, 200:], ref_vals, rtol=1e-3, atol=1) == True
 
 
 def test_transient():
