### 2022-04-10 752dafd
Made run_all executable
