+++ tests/run_all.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import pytest
 import numpy as np
 
