### 2022-04-10 9dca4a9
Fixed some README typos
