### 2022-04-11 #4306 @tpurcell90
There are still a few functions in `controller.py` and all of `neural_network.py` that are undocumented
