+++ OpenPelt/controller.py
@@ -65,8 +65,8 @@ class controller(ABC):
 
         @param t: t is the current timestep.
 
-        @param ref: reference temperature to which the controller steers the respective
-        plate's heat sink
+        @param ref: reference temperature to which the controller steers the
+        respective plate's heat sink
 
         @param sensor_dict: sensor_dict maps "th" or "tc" strings to
         an array of historical values of hot side and cold side temperatures.
@@ -75,22 +75,63 @@ class controller(ABC):
 
 if INCLUDE_TORCH:
     class fake_neural_controller(controller):
-
+        """
+        This class implements a neural network controller. It inherits from the
+        controller class and implements a _controller_f method that calls a Pytorch
+        object (in this case a simple MLP network).
+        """
+    
         def __init__(self, seqr):
+            """
+            Set the sequencer in use. Intended to be an instance of a subclass of
+            sequencer. Furthermore, it initializes the neural controller (neural
+            network).
+    
+            @param seqr: sequencer object in use
+            """
             self.seqr = seqr
-
+    
+            # Initialize an MLP (see file neural_networks.py for more details on
+            # the MLP)
             self.net = MLP(input_units=3,
                            hidden_units=64,
                            output_units=1,
                            bias=True).eval()
-
+    
         def scale(self, x, var=(-100, 100), feature_range=(-1, 1)):
+            """
+            It scales argument x in the interval defines by feature_range.
+    
+            @param x: input scalar to be normalized
+    
+            @param var: A tuple that contains the maximum and the minimum values
+            that x can take
+    
+            @param feature_range: A tuple that containst the desired range of the
+            scaled data
+    
+            @return: The normalized value of x in the interval feature_ranage
+    
+            """
             x_std = (x - var[0]) / (var[1] - var[0])
             x_scaled = (x_std * (feature_range[1] - feature_range[0]) +
                         feature_range[0])
             return x_scaled
-
+    
         def _controller_f(self, t, ref, sensor_dict):
+            """
+            Clamp outputs at 16.00 at the high end and -6.00 at the low end.
+    
+            @param t: t is the current timestep.
+    
+            @param ref: reference temperature to which the controller steers the
+            respective plate's heat sink
+    
+            @param sensor_dict: sensor_dict maps "th" or "tc" strings to
+            an array of historical values of hot side and cold side temperatures.
+    
+            @return: Output value to drive TEC circuit model
+            """
             th = self.scale(sensor_dict['th'][-1])
             tc = self.scale(sensor_dict['tc'][-1])
             ref = self.scale(ref)
@@ -100,18 +141,50 @@ if INCLUDE_TORCH:
 
 
 class random_agent_controller(controller):
+    """
+    This class implements a naive (random) agent. It demonstrates how the user
+    can implement reinforcement learning algorithms using OpenPelt as a TEC
+    simulator.
+    """
 
     def __init__(self, seqr):
+        """
+        Set the sequencer in use. Intended to be an instance of a subclass of
+        sequencer. Furthermore, it sets up all the total reward of the agent
+        (total_r), the set of permited actions the agent can take (volt). The
+        actions are voltage values that pass to the OpenPelt simulator and
+        controll the temperature. The parameter freeze determines if the
+        episode has ended meaenint the agent has succesfully set the
+        temperature to the reference value.
+
+        @param seqr: sequencer object in use
+        """
         self.seqr = seqr
-        self.ref = 0.0
-        self.volt = [-6.0 + i for i in range(23)]
-        self.total_r = 0
-        self.freeze = False
+        self.ref = 0.0          # reference temperature value
[…]
