### 2022-04-19 #4306 @danaraujocr
Hi Rachel,

Sorry about the delay. Had an issue with my computer and just solved it.
Will continue the work later today.

Daniel Rodrigues

PhD Researcher - Process Control / Multivariate Statistics

Université Laval
+1 581 992 0165 <+1+581+992+0165>
***@***.***
401 25-3655 Boulevard des Trinitaires, Montréal, QC, Canada
[image: linkedin] <https://www.linkedin.com/in/daniel-rodrigues-01557668/>

Em seg., 18 de abr. de 2022 às 13:43, Rachel Kurchin <
***@***.***> escreveu:

> 👋 Hi @danaraujocr <https://github.com/danaraujocr>, just checking in on
> how your review is going!
>
