### 2022-04-21 #4306 @danaraujocr
The paper is well written, but the installation doc are still not perfectly clear. I ran a new ubuntu 20.04 installation on a virtual machine to check if the installation would run exactly as described but it didn't for minor reasons. The few things I would suggest adding/changing are:
1 - in the build_nspice.sh file a YACC command is called and those were not available in the build I used. Easily solvable. I commented on the issue opened by @tpurcell90 adding this complication.
2 - In the instructions to add the lbngspice environment variable the text suggests that one needs to copy/paste the snippet into .bashrc and what is in it is a command that does that for you. that might be confusing for some users.
3 - In the snippet that contains the code below :
$ cd OpenPelt/
$ ./build_ngspice.sh
$ pip3 (or pip) install .
I think you should only keep the last line, as the two previous lines have probably already been executed. (especially the shell script)

Other than those minor things, nothing else to add.
