### 2022-04-22 cb8c3d5
README modified based on Reviewer's comments
