### 2022-05-02 #4306 @rkurchin
Some small editorial comments:
* There's a weirdly-formatted citation at the end of the first paragraph in the Statement of Need, where one typesets in square brackets and the other not (**broad note on citation formatting, to cite multiple sources at once, you can do e.g. `[@doe99; @smith2000; @smith2004]`) to get them all in one set of parentheses**)
* line 29: hyphenate "several-minute" (compound adjective)
* 45: hyphenate "open-source" (ditto)
* I suggest surrounding code with `` to get monospace typesetting, e.g. for mentions of `tec_plant` et al.
* line 71: should be "...unlikely to be very many..."
* 74: suggest rephrasing to "...they are solved using iterative methods for systems of differential equations." (removes duplication of "solve" in same sentence)
