+++ paper.md
@@ -34,7 +34,7 @@ Thermoelectric coolers (TECs) are semiconductor heat pumps used in various appli
 For instance, when heat sinking and fans will not suffice in electronics cooling
 applications, a TEC can be used as an alternative. Similarly, TECs can be used
 to cool down components in lasers and volatile organic compound
-detectors [@MansourPaper][@MOSIERBOSS200315].
+detectors [@MansourPaper; @MOSIERBOSS200315].
 
 Designing TEC-based temperature control systems involves overcoming two primary
 challenges. One of these is hardware design. The other is control algorithm
@@ -44,7 +44,7 @@ temperature with a thermoelectric cooler can be on the order of a minute.
 To observe an oscillatory temperature characteristic fully settle can take
 a few additional minutes. In order to fully understand the performance of a
 control algorithm with particular parameters, one may need to endure these
-several minute delays repeatedly. The result is a very arduous process.
+several-minute delays repeatedly. The result is a very arduous process.
 Self-tuning algorithms exist in the literature that partially resolve these
 issues. However, it is wise to prototype any control algorithm in a simulation
 beforehand for faster-than-real-time result acquisition and to avoid potentially
@@ -65,7 +65,7 @@ actuator.
 
 # Object-Oriented SPICE-based Thermoelectric Cooler Model
 
-To our knowledge, no out-of-the-box, open source solution for TEC controller simulation
+To our knowledge, no out-of-the-box, open-source solution for TEC controller simulation
 exists. This is partly due to the lack of a usable open source model of a
 thermoelectric cooler and the lack of open source software to simulate it.
 We investigated the literature and found an electro-thermal circuit model
@@ -73,8 +73,8 @@ of a TEC [@848895]. Though the SPICE netlist is publicly available, there is no 
 approach for simulating a traditional, let alone novel, control algorithm.
 
 We wrote a circuit netlist for the TEC model and buried the implementation
-details in a tec_plant class. We then developed test and controller objects
-that enable rapid testing of control algorithms on instances of this tec_plant class.
+details in a ``tec_plant`` class. We then developed test and controller objects
+that enable rapid testing of control algorithms on instances of this ``tec_plant`` class.
 Firstly, one needs access to an open source SPICE simulator, ngspice in our case.
 The simulator then needs to support external current/voltage sources, so that
 a user's Python function can generate the driving value to the simulator on
@@ -87,7 +87,7 @@ code and use it to interact with the TEC model. We used the PySpice library
 as a wrapper around ngspice to export the simulator's shared library
 functionality to Python [@PySpice].
 
-We default the tec_plant model parameters, such as thermal capacitance
+We default the ``tec_plant`` model parameters, such as thermal capacitance
 and resistance values, to the values in the original paper. However, users
 are able to modify them with their experimental results or theoretical
 approximations. In the future, we would like to provide detailed models of
@@ -100,11 +100,11 @@ to the ideality of the sensor and its coupling to the TEC.
 Furthermore, the model is useful for relatively low-power control of the TEC.
 However, at higher power levels, the TEC heats up and thermal drift affects
 the circuit parameters. This is also a desirable future feature, but there are
-unlikely very many applications in which this is useful.
+unlikely to be very many applications in which this is useful.
 
 We ultimately chose a circuit model for two reasons. Firstly, circuit models
 are computationally simple and efficient since they are solved using
-iterative methods to solve systems of differential equations. However,
[…]
