### 2022-05-03 #4306 @thejackal360
Version Number: beta
DOI: 10.5281/zenodo.6514903

Are we able to change the version to "beta" for the submission? JOSS won't let us use "@editorialbot set beta as version" since we are not editors.
