+++ paper.bib
@@ -74,6 +74,7 @@ Casas, Diego and others},
   number={7897},
   pages={414--419},
   year={2022},
+  doi={10.1038/s41586-021-04301-9},
   publisher={Nature Publishing Group}
 }
 
